# internal/dto/types.gen.go
// Package dto provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package dto

import (
	"time"
)

// Defines values for ErrorResponseErrorCode.
const (
	NOCANDIDATE ErrorResponseErrorCode = "NO_CANDIDATE"
	NOTASSIGNED ErrorResponseErrorCode = "NOT_ASSIGNED"
	NOTFOUND    ErrorResponseErrorCode = "NOT_FOUND"
	PREXISTS    ErrorResponseErrorCode = "PR_EXISTS"
	PRMERGED    ErrorResponseErrorCode = "PR_MERGED"
	TEAMEXISTS  ErrorResponseErrorCode = "TEAM_EXISTS"
)

// Defines values for PullRequestStatus.
const (
	PullRequestStatusMERGED PullRequestStatus = "MERGED"
	PullRequestStatusOPEN   PullRequestStatus = "OPEN"
)

// Defines values for PullRequestShortStatus.
const (
	PullRequestShortStatusMERGED PullRequestShortStatus = "MERGED"
	PullRequestShortStatusOPEN   PullRequestShortStatus = "OPEN"
)

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Error struct {
		Code    ErrorResponseErrorCode `json:"code"`
		Message string                 `json:"message"`
	} `json:"error"`
}

// ErrorResponseErrorCode defines model for ErrorResponse.Error.Code.
type ErrorResponseErrorCode string

// PullRequest defines model for PullRequest.
type PullRequest struct {
	// AssignedReviewers user_id назначенных ревьюверов (0..2)
	AssignedReviewers []string          `json:"assigned_reviewers"`
	AuthorId          string            `json:"author_id"`
	CreatedAt         *time.Time        `json:"createdAt"`
	MergedAt          *time.Time        `json:"mergedAt"`
	PullRequestId     string            `json:"pull_request_id"`
	PullRequestName   string            `json:"pull_request_name"`
	Status            PullRequestStatus `json:"status"`
}

// PullRequestStatus defines model for PullRequest.Status.
type PullRequestStatus string

// PullRequestShort defines model for PullRequestShort.
type PullRequestShort struct {
	AuthorId        string                 `json:"author_id"`
	PullRequestId   string                 `json:"pull_request_id"`
	PullRequestName string                 `json:"pull_request_name"`
	Status          PullRequestShortStatus `json:"status"`
}

// PullRequestShortStatus defines model for PullRequestShort.Status.
type PullRequestShortStatus string

// Team defines model for Team.
type Team struct {
	Members  []TeamMember `json:"members"`
	TeamName string       `json:"team_name"`
}

// TeamMember defines model for TeamMember.
type TeamMember struct {
	IsActive bool   `json:"is_active"`
	UserId   string `json:"user_id"`
	Username string `json:"username"`
}

// User defines model for User.
type User struct {
	IsActive bool   `json:"is_active"`
	TeamName string `json:"team_name"`
	UserId   string `json:"user_id"`
	Username string `json:"username"`
}

// TeamNameQuery defines model for TeamNameQuery.
type TeamNameQuery = string

// UserIdQuery defines model for UserIdQuery.
type UserIdQuery = string

// PostPullRequestCreateJSONBody defines parameters for PostPullRequestCreate.
type PostPullRequestCreateJSONBody struct {
	AuthorId        string `json:"author_id"`
	PullRequestId   string `json:"pull_request_id"`
	PullRequestName string `json:"pull_request_name"`
}

// PostPullRequestMergeJSONBody defines parameters for PostPullRequestMerge.
type PostPullRequestMergeJSONBody struct {
	PullRequestId string `json:"pull_request_id"`
}

// PostPullRequestReassignJSONBody defines parameters for PostPullRequestReassign.
type PostPullRequestReassignJSONBody struct {
	OldUserId     string `json:"old_user_id"`
	PullRequestId string `json:"pull_request_id"`
}

// GetTeamGetParams defines parameters for GetTeamGet.
type GetTeamGetParams struct {
	// TeamName Уникальное имя команды
	TeamName TeamNameQuery `form:"team_name" json:"team_name"`
}

// GetUsersGetReviewParams defines parameters for GetUsersGetReview.
type GetUsersGetReviewParams struct {
	// UserId Идентификатор пользователя
	UserId UserIdQuery `form:"user_id" json:"user_id"`
}

// PostUsersSetIsActiveJSONBody defines parameters for PostUsersSetIsActive.
type PostUsersSetIsActiveJSONBody struct {
	IsActive bool   `json:"is_active"`
	UserId   string `json:"user_id"`
}

// PostPullRequestCreateJSONRequestBody defines body for PostPullRequestCreate for application/json ContentType.
type PostPullRequestCreateJSONRequestBody PostPullRequestCreateJSONBody

// PostPullRequestMergeJSONRequestBody defines body for PostPullRequestMerge for application/json ContentType.
type PostPullRequestMergeJSONRequestBody PostPullRequestMergeJSONBody

// PostPullRequestReassignJSONRequestBody defines body for PostPullRequestReassign for application/json ContentType.
type PostPullRequestReassignJSONRequestBody PostPullRequestReassignJSONBody

// PostTeamAddJSONRequestBody defines body for PostTeamAdd for application/json ContentType.
type PostTeamAddJSONRequestBody = Team

// PostUsersSetIsActiveJSONRequestBody defines body for PostUsersSetIsActive for application/json ContentType.
type PostUsersSetIsActiveJSONRequestBody PostUsersSetIsActiveJSONBody