# internal/entity/team.go
package entity

type Team struct {
    Name    string       `json:"team_name"`
    Members []TeamMember `json:"members"`
}

type TeamMember struct {
    UserID   string `json:"user_id"`
    Username string `json:"username"`
    IsActive bool   `json:"is_active"`
}